Use Windows TZID names
The TZID name is arbitrary - it's just a way for a VEVENT to reference
the particular VTIMEZONE event for that event's time zone. The event
will therefore show the correct time adjusted for timezone regardless of
the name of the TZID. However, if Windows-style TZID names are used the
.ics better supports Outlook: the event view will show the original time
with the actual time zone instead of the adjusted time with the current
time zone.